remove visual studio project files.
the build process for the library now depends on auto-generated files which
don't get created when building straight from the existing solution.

it's possible to create a new solution/project file with CMake which will
check the avilability of certain platform-specific features and create a
header file with #define's for them. this is currently only used to detect
wether the target platform provides a 'stdint.h' header but can be extended
to just about any other header/library in the future.